Split by file in 6 different files
