+++ pages/crash_stats_nightly_builds_page.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python
+
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+from selenium.webdriver.common.by import By
+
+from pages.base_page import CrashStatsBasePage
+
+
+class CrashStatsNightlyBuilds(CrashStatsBasePage):
+
+    _link_to_ftp_locator = (By.CSS_SELECTOR, '.notitle > p > a')
+
+    @property
+    def link_to_ftp(self):
+        return self.selenium.find_element(*self._link_to_ftp_locator).get_attribute('href')
+
+    def click_link_to_ftp(self):
+        self.selenium.find_element(*self._link_to_ftp_locator).click()
