+++ pages/crash_stats_top_changers_page.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+from pages.base_page import CrashStatsBasePage
+
+
+class CrashStatsTopChangers(CrashStatsBasePage):
+
+    pass
