+++ pages/signature_report_page.py
@@ -1,0 +1,24 @@
+#!/usr/bin/env python
+
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+from selenium.webdriver.common.by import By
+
+from pages.base_page import CrashStatsBasePage
+
+
+class SignatureReport(CrashStatsBasePage):
+
+    # https://crash-stats.allizom.org/report/list?
+
+    _total_items_locator = (By.CSS_SELECTOR, 'span.totalItems')
+    _reports_page_locator = (By.CSS_SELECTOR, '.ui-state-default.ui-corner-top:nth-of-type(4) > a > span')
+
+    def click_reports(self):
+        self.selenium.find_element(*self._reports_page_locator).click()
+
+    @property
+    def total_items_label(self):
+        return self.selenium.find_element(*self._total_items_locator).text.replace(",", "")
